Adds diagram stubs, and removes the bin/clusternator-es5.js, adds postinstall hooks to transpile
