+++ bin/clusternator.sh
@@ -8,6 +8,6 @@ while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symli
 done
 
 DIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
-SCRIPT=/clusternatorCli-es5.js
+SCRIPT=/clusternatorCli.js
 
 node $DIR$SCRIPT $@
