+++ gulpfile.js
@@ -1,18 +1,27 @@
 'use strict';
 
 var gulp = require('gulp'),
-babel = require('gulp-babel'),
-eslint = require('gulp-eslint'),
-mocha = require('gulp-mocha'),
-istanbul = require('gulp-istanbul');
+  babel = require('gulp-babel'),
+  eslint = require('gulp-eslint'),
+  mocha = require('gulp-mocha'),
+  istanbul = require('gulp-istanbul');
 
 var jsPaths = ['src/**/*.js'],
-specPaths = ['spec/unit/**/*.spec.js'];
+  cliPath = ['bin-src/**/*.js'],
+  specPaths = ['spec/unit/**/*.spec.js'];
 
 gulp.task('default', ['transpile']);
 gulp.task('test', ['test-unit']);
 
-gulp.task('transpile', function transpile() {
+gulp.task('transpile', ['transpile-cli', 'transpile-src']);
+
+gulp.task('transpile-cli', function transpileCli() {
+  return gulp.src(cliPath).
+  pipe(babel()).
+  pipe(gulp.dest('bin'));
+});
+
+gulp.task('transpile-src', function transpileSrc() {
   return gulp.src(jsPaths).
   pipe(babel()).
   pipe(gulp.dest('lib'));
@@ -28,17 +37,24 @@ gulp.task('pre-test-unit', ['lint'], function preUnitTest() {
 });
 
 gulp.task('test-unit', ['pre-test-unit'], function testUnit() {
-    return gulp.src(specPaths).
-    pipe(mocha()).
-    pipe(istanbul.writeReports({
-      reporters: [ 'text', 'lcovonly', 'html', 'json', 'text-summary'],
-      reportOpts: {
-        dir: './coverage',
-        lcov: { dir: 'coverage/lcovonly', file: 'lcov.info' },
-        html: { dir: 'coverage/html' },
-        json: { dir: 'coverage/json' }
+  return gulp.src(specPaths).
+  pipe(mocha()).
+  pipe(istanbul.writeReports({
+    reporters: ['text', 'lcovonly', 'html', 'json', 'text-summary'],
+    reportOpts: {
+      dir: './coverage',
+      lcov: {
+        dir: 'coverage/lcovonly',
+        file: 'lcov.info'
+      },
+      html: {
+        dir: 'coverage/html'
+      },
+      json: {
+        dir: 'coverage/json'
       }
-    }));
+    }
+  }));
 });
 
 gulp.task('lint', function lint() {
